+++ python/mocpy/abstract_moc.py
@@ -51,6 +51,12 @@ class AbstractMOC(serializer.IO, metaclass=abc.ABCMeta):
     def __deepcopy__(self, memo):
         return self.__copy__()
 
+    def __setstate__(self, state):
+        # this is called when a MOC is unpickled
+        # we create a new ref count with copy
+        mocpy.copy(state["store_index"])
+        self.__dict__ = state
+
     @staticmethod
     def _store_index_dtype():
         """Store the datatype of the index.