+++ CHANGELOG.md
@@ -16,6 +16,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [unreleased]
 
+## [0.14.0]
+
 ### Fixed
 
 * `MOC.border` no does not attempt on plotting the border when the MOc is out of the